add hook useRestore
