+++ src/screens/Restore/Mnemonic.tsx
@@ -7,7 +7,7 @@ import type { IMnemonicPageProps } from '@model/nav'
 import { useThemeContext } from '@src/context/Theme'
 import { NS } from '@src/i18n'
 import { mainColors } from '@styles'
-import { generateMnemonic } from '@wallet/restore'
+import { generateMnemonic } from '@wallet'
 import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { FlatList, SafeAreaView, View } from 'react-native'
